image/test/mochitest/test_bug552605-1.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=552605
-->
<head>
  <title>Test for Bug 552605</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/WindowSnapshot.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=552605">Mozilla Bug 552605</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

var first, second;

SimpleTest.waitForExplicitFinish();

function checkFirst()
{
  var testimage = document.getElementById('test-image');
  first = document.createElement('canvas')
  var ctx = first.getContext('2d');
  ctx.drawImage(testimage, 0, 0);

  var newimg = new Image();
  newimg.onload = checkSecond;
  newimg.src = "bug552605.sjs";
  document.body.appendChild(newimg);
}

function checkSecond()
{
  var testimage = document.getElementById('test-image');
  second = document.createElement('canvas')
  var ctx = second.getContext('2d');
  ctx.drawImage(testimage, 0, 0);

  // Check that the images are the same, since they're in the same document.
  [correct, val1, val2] = compareSnapshots(first, second, true);
  ok(correct, "Image should be the same for all loads.");

  SimpleTest.finish();
}

</script>
</pre>
<div id="content"> <!-- style="display: none" -->
<img src="bug552605.sjs" onload="checkFirst()" id="test-image"></iframe>
</div>
</body>
</html>
